+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changes for lovely.esdb
 unreleased
 ==========
 
+ - added Document.get_by to build simple terms queries
+ - added WITH_INHERITANCE property for the Document class to enable/disable
+   the use of the db_class__ property.
+
 2016/07/08 0.2.3
 ================